pyproject.toml
[tool.poetry]
name = "desuko"
version = "0.1.0"
description = "Discord bot."
authors = ["Arisu Wonderland <arisuchr@riseup.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
py-cord = "^2.2.2"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
pylint = "^2.15.0"

[tool.isort]
line_length=100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
